model/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
keras
scikit-learn
opencv-python
matplotlib
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv